www/code-server: avoid network access outside the 'fetch' phase (#10)
This change avoids network access outside the 'fetch' phase of
the port build - a FreeBSD best practice and a must if the port
is to be accepted into the ports tree.

This is mostly achieved by bundling node_modules/ in a tarball
that is downloaded and extracted during the build. We are also
cloning the code-server repository and not fetching the source
through npm anymore.