www/code-server: fix port
https://github.com/0x1eef/myports/pull/10 was a step in the right
direction but it didn't cover everything, and the package that was
built would install a broken copy of code-server.

I decided to take a different approach. We now have the build/
subdirectory with its own Makefile. It will produce a tarball that
includes a fully built code-server. This would be done by a maintainer,
and the tarball uploaded as a github reelease.

The port Makefile in the parent directory then only downloads and
extracts the contents of the earlier build.

properly build vscode and code-server together. Beforehand - certain
build artifacts were missing from the tarballs.